Added numeric type expansion to long and double.

+++ src/main/java/net/ninjacat/objmatcher/matcher/patterns/IntegerEq.java
@@ -5,13 +5,13 @@ import lombok.Value;
 
 @Value
 @EqualsAndHashCode(callSuper = true)
-public class IntegerEq extends FieldPattern<Integer> {
-    IntegerEq(final String fieldName, final Integer value) {
-        super(fieldName, Integer.class, value);
+public class IntegerEq extends FieldPattern<Long> {
+    IntegerEq(final String fieldName, final Long value) {
+        super(fieldName, Long.class, value);
     }
 
     @Override
-    public boolean matches(final Integer checkedValue) {
+    public boolean matches(final Long checkedValue) {
         return getValue().compareTo(checkedValue) == 0;
     }
 }
