+++ src/main/java/net/ninjacat/objmatcher/matcher/patterns/IntegerGt.java
@@ -5,14 +5,14 @@ import lombok.Value;
 
 @Value
 @EqualsAndHashCode(callSuper = true)
-public class IntegerGt extends FieldPattern<Integer> {
+public class IntegerGt extends FieldPattern<Long> {
 
-    IntegerGt(final String fieldName, final Integer value) {
-        super(fieldName, Integer.class, value);
+    IntegerGt(final String fieldName, final Long value) {
+        super(fieldName, Long.class, value);
     }
 
     @Override
-    public boolean matches(final Integer checkedValue) {
+    public boolean matches(final Long checkedValue) {
         return checkedValue.compareTo(getValue()) > 0;
     }
 }
