+++ src/main/java/net/ninjacat/objmatcher/matcher/patterns/IntegerLt.java
@@ -5,13 +5,13 @@ import lombok.Value;
 
 @Value
 @EqualsAndHashCode(callSuper = true)
-public class IntegerLt extends FieldPattern<Integer> {
-    IntegerLt(final String fieldName, final Integer value) {
-        super(fieldName, Integer.class, value);
+public class IntegerLt extends FieldPattern<Long> {
+    IntegerLt(final String fieldName, final Long value) {
+        super(fieldName, Long.class, value);
     }
 
     @Override
-    public boolean matches(final Integer checkedValue) {
+    public boolean matches(final Long checkedValue) {
         return checkedValue.compareTo(getValue()) < 0;
     }
 }
